JCR-2112: Moved the unit test into a package allowing it to be launched and replaced special characters with unicode characters to avoid having a file UTF-8 encoded
